+++ packages/koot-boilerplate/src/routes/index.js
@@ -23,9 +23,12 @@ export default {
          * - 推荐使用 koot 提供的 `routeCheck()` 方法保证路由组件的正确渲染 (详情请查阅文档)
          */
         getComponent: (nextState, cb) => {
-            require.ensure([], (require) => {
-                if (routeCheck(nextState)) cb(null, require('@views/home').default)
-            }, 'Page-Home')
+            import(
+                /* webpackChunkName: "PageHome" */
+                '@views/home'
+            ).then(module => {
+                if (routeCheck(nextState)) cb(null, module.default);
+            })
         }
     },
 
@@ -33,9 +36,12 @@ export default {
         {
             path: 'start',
             getComponent: (nextState, cb) => {
-                require.ensure([], (require) => {
-                    if (routeCheck(nextState)) cb(null, require('@views/start').default)
-                }, 'pageStart')
+                import(
+                    /* webpackChunkName: "PageStart" */
+                    '@views/start'
+                ).then(module => {
+                    if (routeCheck(nextState)) cb(null, module.default);
+                })
             }
         },
     ]